requirements.txt
# FastAPI and ASGI server for HTTP endpoints and production
fastapi
uvicorn[standard]
gunicorn

# Firebase and Google Cloud Core
firebase-functions
firebase-admin
google-generativeai
google-generativeai

# Google Cloud Services (Docs API, Secret Manager)
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-cloud-secret-manager

# Vector Database Client
pinecone-client

# Document Parsing
pypdf
python-docx

# HTTP Requests (for Perplexity API and Firebase Auth)
requests

# Environment variables for local development
python-dotenv

# Authentication (JWT)
python-jose[cryptography]
passlib[bcrypt]